Shaders: more cases of accidental lowp.
Should have gone into b93de590bcc778763b0fa5cf911d1e2e9209112e already.
No idea what am I doing there.

+++ src/Magnum/Shaders/Line.in.vert
@@ -45,7 +45,7 @@ highp vec2 expandLineVertex(
     in mediump const float width,
     in mediump const float smoothness,
     in highp const float miterLimit,
-    in lowp const vec2 viewportSize,
+    in highp const vec2 viewportSize,
     out highp vec2 centerDistanceSigned,
     out highp float halfSegmentLength,
     out highp float hasCap)
